IMPLEMENTED Path Planner class which integrates all other components for path planning
PathPlanner\AStar.cs, ISearchProblem.cs: state cost is double which is easier to use for distances
PathPlanner\OccupancyGridPathSearchProblem.cs: start and goal removed from ctor cause they have different life scope
PathPlanner\PathPlanner.cs: uses MapInflater, OccupancyGridPathSearchProblem, JumpPointGenerator and AStar to search for the path




git-svn-id: svn://brumba.ru:46104/brumba@183 cf9289f8-4b28-4085-b22d-9969776bd1f6

+++ src/Brumba/PathPlanner.Tests/AStarTests.cs
@@ -68,8 +68,8 @@ namespace Brumba.PathPlanner.Tests
 		public Func<int, List<int>> ExpandF { get; set; }
 		public Func<int, int> HeuristicF { get; set; }
 
-		public IEnumerable<Tuple<int, int>> Expand(int state) { return ExpandF(state).Select(s => Tuple.Create(s, 1)); }
+		public IEnumerable<Tuple<int, double>> Expand(int state) { return ExpandF(state).Select(s => Tuple.Create(s, 1.0)); }
 
-		public int GetHeuristic(int state) { return HeuristicF(state); }
+		public double GetHeuristic(int state) { return HeuristicF(state); }
 	}
 }
